Sandbox popup (#142)
* Warning for disabling sandbox mode

* submit non workint

* add reset function and modal on existing_cluster page

* fix issue of destroy button

* moved form andremogved reset

* fix

* fix mis-aligned button

* fix cancel button of modal send unexpect request to backend

* print output in iFrame

* Closes #126

* Closes #127

+++ terraform/appflask.py
@@ -21,7 +21,8 @@ from jinja2 import Template
 import argparse
 
 VALID_FABRIC_TYPE = ['aci', 'vxlan_evpn']
-
+TF_STATE_ACI = "terraform.tfstate"
+TF_STATE_NDFC = "./ndfc/terraform.tfstate"
 # app = Flask(__name__)
 app = Flask(__name__, template_folder='./TEMPLATES/')
 app.config['SECRET_KEY'] = 'cisco'
@@ -421,7 +422,7 @@ def create():
         req = request.form
         button = req.get("button")
         if button == "Previous":
-            return redirect('/cluster?fabric_type=' + fabric_type)
+            return redirect('/cluster_network?fabric_type=' + fabric_type)
         if button == "Update Config":
             config = req.get('config')
             with open('cluster.tfvars', 'w', encoding='utf-8') as f:
@@ -1380,6 +1381,31 @@ def read_process(g):
             yield line
 
 
+@app.route('/reset', methods=['GET'])
+@require_api_token
+def reset():
+    '''generic function to delete the TF State'''
+    fabric_type = get_fabric_type(request)
+
+    if fabric_type not in VALID_FABRIC_TYPE:
+        return redirect('/')
+    if request.method == "GET":
+        try:
+            if fabric_type == "aci":
+                if os.path.exists(TF_STATE_ACI):
+                    os.remove(TF_STATE_ACI)
+                    return Response("Deleted terraform state file " + TF_STATE_ACI)
+                else:
+                    return Response("terraform state file " + TF_STATE_ACI +" Not found")
+            if fabric_type == "vxlan_evpn":
+                if os.path.exists(TF_STATE_NDFC):
+                    os.remove(TF_STATE_NDFC)
+                    return Response("Deleted terraform state file " + TF_STATE_NDFC)
+                else:
+                    return Response("terraform state file " + TF_STATE_NDFC +" Not found")
+        except Exception:
+           return Response("Reset Failed")
+
 @app.route('/existing_cluster', methods=['GET', 'POST'])
 @require_api_token
 def existing_cluster():
@@ -1388,16 +1414,20 @@ def existing_cluster():
     if fabric_type not in VALID_FABRIC_TYPE:
         return redirect('/')
     if request.method == "POST":
-        g = proc.Group()
-        if fabric_type == "aci":
-            g.run(["bash", "-c", "terraform destroy -auto-approve -no-color -var-file='cluster.tfvars' && \
-            ansible-playbook -i ../ansible/inventory/apic.yaml ../ansible/apic_user.yaml --tags='apic_user_del'"])
-        elif fabric_type == "vxlan_evpn":
-            g.run(["bash",
-                   "-c",
-                   "terraform -chdir ndfc destroy -auto-approve -no-color -var-file='cluster.tfvars'"])
-        #p = g.run("ls")
-        return Response(read_process(g), mimetype='text/event-stream')
+        req = request.form
+        button = req.get("button")
+        print(button)
+        if button == "destroy":
+            g = proc.Group()
+            if fabric_type == "aci":
+                g.run(["bash", "-c", "terraform destroy -auto-approve -no-color -var-file='cluster.tfvars' && \
+                ansible-playbook -i ../ansible/inventory/apic.yaml ../ansible/apic_user.yaml --tags='apic_user_del'"])
+            elif fabric_type == "vxlan_evpn":
+                g.run(["bash",
+                    "-c",
+                    "terraform -chdir ndfc destroy -auto-approve -no-color -var-file='cluster.tfvars'"])
+            #p = g.run("ls")
+            return Response(read_process(g), mimetype='text/event-stream')
     else:
 
         if fabric_type == "aci":
@@ -1462,19 +1492,17 @@ def get_page():
         if button == "Next":
             return redirect(f'/login?fabric_type={fabric_type}')
     if request.method == "GET":
-        tf_state_aci = "terraform.tfstate"
-        tf_state_ndfc = "./ndfc/terraform.tfstate"
         # if no tf state existed, return the intro page
-        if not os.path.exists(tf_state_aci) and not os.path.exists(tf_state_ndfc):
+        if not os.path.exists(TF_STATE_ACI) and not os.path.exists(TF_STATE_NDFC):
             return render_template('intro.html')
 
-        if os.path.exists(tf_state_aci):
+        if os.path.exists(TF_STATE_ACI):
             with open('terraform.tfstate', 'r', encoding='utf-8') as f:
                 state_aci = json.load(f)
             if state_aci['resources'] != []:
                 return redirect('/existing_cluster')
 
-        if os.path.exists(tf_state_ndfc):
+        if os.path.exists(TF_STATE_NDFC):
             with open('./ndfc/terraform.tfstate', 'r', encoding='utf-8') as f:
                 state_ndfc = json.load(f)
                 # If there are resources the cluster is there
